% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.uspto.data.R
\name{download.uspto.data}
\alias{download.uspto.data}
\title{download.uspto.data}
\usage{
download.uspto.data(data.type = NA, file.type = NA, sample = NA)
}
\arguments{
\item{data.type}{Specify the data type you want to download. For bibliographic the argument is "bibl". For full text data the argument is "full".}

\item{file.type}{Specify the type of file you want to download. The arguments for bibliographic data are "ipgb", "pgb", and "pba". The arguments for Full text data are "pftaps", "pg", and "ipg".}

\item{sample}{Specify the years from which you want to download the data. Note that you need to use "parentheses".}
}
\description{
The function download.uspto.data downloads bibliographic or full text patent data from the USPTO database.
Bibliographic data contains front page information from each patent that was issued by the uspto from 1976 until present. Full text data contains the full text of each patent grant issued weekly from 1976 to present.
The function creates directories in your home directory which correspond to the data types and file types. Before downloading, the function will check whether the required data is allready in the folder, if this is not the case,
the data will be downloaded and stored in the corresponding folder.
}
\details{
When entering the sample, parentheses need to included; in example "2016". When the sample consists of multiple years specify the argument like this: "2015|2016|2017".

The function makes use of a configuration file with a yaml extention. This configuration file needs to be stored in the home directory and needs to have this specific name: "intera.config.yml".
The configuration file contains user specific paths to specific directories where the data are stored. The configuration file consists of keys and values. Each line of the configuration file contains "key: value". The key represents an object in R (which is the name of a directory), the value is the path to this directory.
}

